entrega_2/requirements.txt
asn1crypto==1.5.1
asttokens==2.2.1      
backcall==0.2.0       
beautifulsoup4==4.12.2
boto3==1.26.153       
botocore==1.29.153    
certifi==2023.5.7
charset-normalizer==3.1.0
colorama==0.4.6
comm==0.1.3
debugpy==1.6.7
decorator==5.1.1
executing==1.2.0
greenlet==2.0.2
idna==3.4
ipykernel==6.23.3
ipython==8.14.0
jedi==0.18.2
jmespath==1.0.1
jupyter_client==8.3.0
jupyter_core==5.3.1
lxml==4.9.2
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
numpy==1.24.3
packaging==23.1
pandas==2.0.2
parso==0.8.3
pickleshare==0.7.5
platformdirs==3.8.0
prompt-toolkit==3.0.38
psutil==5.9.5
psycopg2==2.9.6
pure-eval==0.2.2
py4j==0.10.9.7
Pygments==2.15.1
pyspark==3.4.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
pywin32==306
pyzmq==25.1.0
redshift-connector==2.0.911
requests==2.31.0
s3transfer==0.6.1
scramp==1.4.4
six==1.16.0
soupsieve==2.4.1
SQLAlchemy==1.4.48
sqlalchemy-redshift==0.8.14
stack-data==0.6.2
tornado==6.3.2
traitlets==5.9.0
typing_extensions==4.7.0
tzdata==2023.3
urllib3==1.26.16
wcwidth==0.2.6